refactor: move message dispatch from service to handler
