+++ biz/service/chat/send.go
@@ -1,6 +1,8 @@
 package service
 
-func (s *ChatService) sendMessageToUser(userID int64, msgType int, body any) (bool, error) {
+import "github.com/ACaiCat/tiktok-go/pkg/errno"
+
+func (s *ChatService) sendMessage(userID int64, msgType int, body any) (bool, error) {
 	user, online := s.manager.GetOnlineUser(userID)
 	if !online {
 		return false, nil
@@ -11,4 +13,10 @@ func (s *ChatService) sendMessageToUser(userID int64, msgType int, body any) (bo
 	}
 
 	return true, nil
+}
+
+func (s *ChatService) SendErr(userID int64, err errno.ErrNo) {
+	if u, online := s.manager.GetOnlineUser(userID); online {
+		u.SendError(int(err.ErrCode), err.ErrMsg)
+	}
 }
